chore: refactoring extractions
