+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "upscaler-cli",
-  "version": "1.2.0-dev",
+  "version": "1.3.0-dev",
   "description": "upscale png images from your terminal",
   "main": "./dist/index.js",
   "scripts": {
